backend/requirements.txt
fastapi==0.104.1
motor==3.3.2
pydantic==1.10.2
python-dotenv==1.0.0
